back/requirements.txt
django
yt-dlp
pygame
daphne
ytmusicapi
django-cleanup
Twisted[tls,http2]
django-ninja
orjson
mutagen
